Extract interface from SOAPHelper
This should make it easier to enforce proper layering and avoid cycles.

Also move the implementation classes to avoid a cycle.
